Do not set Connection.Request to null, and add HasHandler atomic
Setting Connection.Request to null is premature, and will actually
cause access violations in multi-threaded servers. There is no need
to do this - the Request object will get cleaned up when the
Connection object is destroyed.

Added HasHandler, which makes it easy for user code to ensure that
only a single thread takes ownership of a request.