Allow sending response before request body is finished
This is useful when dealing with uploads. Often, you'll check the
headers, and determine that this upload is going to fail
(eg authentication invalid), so you'd rather fail it immediately than
waiting for the user to finish the upload before saying "not allowed".

The HTTP spec implies that one should be able to do this, and so we
now allow it. It makes client code simpler.

Also, fixed the support of chunked receives, which was never actually
tested.